refactor: rearrange component files
