+++ src/components/index.js
@@ -1,0 +1,13 @@
+export { default as Navbar } from "./common/Navbar";
+export { default as SearchDropdown } from "./common/SearchDropdown";
+export { default as Sidebar } from "./common/Sidebar";
+export { default as PostCard } from "./community/PostCard";
+export { default as CommunityCard } from "./community/CommunityCard";
+export { default as CommentCard } from "./community/CommentCard";
+export { default as ProductCard } from "./market/ProductCard";
+export { default as PostHistoryCard } from "./mypage/community/PostHistoryCard";
+export { default as CommentHistoryCard } from "./mypage/community/CommentHistoryCard";
+export { default as PurchaseHistoryContainer } from "./mypage/market/PurchaseHistoryCard";
+export { default as SellHistoryContainer } from "./mypage/market/SellHistoryContainer";
+export { default as InboxContainer } from "./mypage/message/InboxContainer.js";
+export { default as OutboxContainer } from "./mypage/message/OutboxContainer";
