+++ src/pages/MarketHistory/index.js
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import Styled from "styled-components";
 import SellHistoryContainer from "../../components/mypage/market/SellHistoryContainer";
 import PurchaseHistoryContainer from "../../components/mypage/market/PurchaseHistoryContainer";
-import Navbar from "../../components/Navbar";
+import { Navbar } from "../../components";
 
 const MarketHistoryWrapper = Styled.div`
   margin: 110px;
