+++ src/pages/MessageHistory/index.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import Styled from "styled-components";
 import InboxContainer from "../../components/mypage/message/InboxContainer";
-import Navbar from "../../components/Navbar";
+import { Navbar } from "../../components";
 import OutboxContainer from "../../components/mypage/message/OutboxContainer";
 
 const MessageHistoryWrapper = Styled.div`
