+++ src/pages/NewPostWrite/index.js
@@ -5,8 +5,8 @@ import InputLabel from "@material-ui/core/InputLabel";
 import MenuItem from "@material-ui/core/MenuItem";
 import FormControl from "@material-ui/core/FormControl";
 import Select from "@material-ui/core/Select";
-import SearchDropdown from "../../components/SearchDropdown";
-import Navbar from "../../components/Navbar";
+import { SearchDropdown } from "../../components";
+import { Navbar } from "../../components";
 import { writePost } from "../../lib/api/post";
 import { withRouter } from "react-router-dom";
 import { getAladinBooks } from "../../lib/api/aladin";
