+++ src/pages/PostDetail/index.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import Styled from "styled-components";
 import CommentCard from "../../components/community/CommentCard";
-import Navbar from "../../components/Navbar";
+import { Navbar } from "../../components";
 import { getPost } from "../../lib/api/post";
 import { writeComment } from "../../lib/api/comment";
