+++ src/pages/PostHistory/index.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import Styled from "styled-components";
 import PostHistoryCard from "../../components/mypage/community/PostHistoryCard";
-import Navbar from "../../components/Navbar";
+import { Navbar } from "../../components";
 import { getPostHistory } from "../../lib/api/user";
 
 const PostHistoryWrapper = Styled.div`
