+++ src/pages/PostWrite/index.js
@@ -5,7 +5,7 @@ import InputLabel from "@material-ui/core/InputLabel";
 import MenuItem from "@material-ui/core/MenuItem";
 import FormControl from "@material-ui/core/FormControl";
 import Select from "@material-ui/core/Select";
-import Navbar from "../../components/Navbar";
+import { Navbar } from "../../components";
 import { updatePost, writePost } from "../../lib/api/post";
 import { withRouter } from "react-router-dom";
